[REFACOTR] Update MyPage Modal Style and divide
마이페이지 이동 모달 스타일 수정 및 분리
